requirements.txt
asgiref==3.4.1
Django==3.2.8
django-extensions==3.1.3
django-filter==21.1
djangorestframework==3.12.4
mysqlclient==2.0.3
numpy==1.21.2
opencv-python==4.5.3.56
pdf2image==1.16.0
Pillow==8.4.0
pyasn1==0.4.8
pysmb==1.2.7
python-magic==0.4.24
pytz==2021.3
sqlparse==0.4.2
